fix: 유효하지 않은 토큰에 대해 프로필 정보 리패치

+++ frontend/src/services/requests/profile.ts
@@ -3,7 +3,11 @@ import axios from "axios";
 import { Profile } from "../../@types";
 import { API_URL } from "../../constants/urls";
 
-export const requestGetSelfProfile = async (accessToken: string) => {
+export const requestGetSelfProfile = async (accessToken: string | null) => {
+  if (!accessToken) {
+    throw Error("no accessToken");
+  }
+
   const response = await axios.get<Profile>(API_URL.SELF_PROFILE, {
     headers: {
       Authorization: `Bearer ${accessToken}`,
@@ -26,7 +30,11 @@ export const requestGetUserProfile = async (username: string, accessToken: strin
   return response.data;
 };
 
-export const requestAddFollow = async (username: string, accessToken: string) => {
+export const requestAddFollow = async (username: string | undefined, accessToken: string | null) => {
+  if (!accessToken || !username) {
+    throw Error("Invalid Request");
+  }
+
   const response = await axios.post(API_URL.USER_PROFILE_FOLLOW(username), null, {
     headers: {
       Authorization: `Bearer ${accessToken}`,
@@ -36,7 +44,11 @@ export const requestAddFollow = async (username: string, accessToken: string) =>
   return response.data;
 };
 
-export const requestDeleteFollow = async (username: string, accessToken: string) => {
+export const requestDeleteFollow = async (username: string | undefined, accessToken: string | null) => {
+  if (!accessToken || !username) {
+    throw Error("Invalid Request");
+  }
+
   const response = await axios.delete(API_URL.USER_PROFILE_FOLLOW(username), {
     headers: {
       Authorization: `Bearer ${accessToken}`,
